+++ app/modules/common/factories/CustomAPIExtension.js
@@ -64,7 +64,15 @@ angular.module('common')
         	type: 'GET',
         	accepts: 'application/vnd.com.covisint.platform.package.grant.v1+json',
         	call: '/registration/v1/registrations/organizations/{organizationId}/packages'
-        }
+        },
+		{
+			cmd: 'validateUsernameEmailNonce',
+			cmdType: 'nonce',
+			type: 'POST',
+			accepts: 'application/vnd.com.covisint.platform.person.password.account.v1+json',
+			contentType: 'application/vnd.com.covisint.platform.person.password.account.v1+json',
+			call: '/registration/v1/registrations/persons/registration/validate'
+		}
     ];
 
     return calls